requirements_static.txt
absl-py==1.2.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
astunparse==1.6.3
backcall==0.2.0
bleach==5.0.1
cached-property==1.5.2
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
chex==0.1.5
clip @ git+https://github.com/openai/CLIP.git@3702849800aa56e2223035bccd1c6ef91c704ca8
cloudpickle==2.2.0
clu==0.0.8
colorama==0.4.5
commonmark==0.9.1
contextlib2==21.6.0
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
dm-reverb-nightly==0.11.0.dev20230205
dm-tree==0.1.7
entrypoints==0.4
etils==0.6.0
executing==1.2.0
fastjsonschema==2.16.2
flatbuffers==22.10.26
flax==0.5.3
fonttools==4.38.0
ftfy==6.1.1
gast==0.4.0
gin-config==0.5.0
google-api-core==2.10.1
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-cloud-speech==2.16.0
google-pasta==0.2.0
googleapis-common-protos==1.56.4
grain==0.1.4
grpcio==1.49.1
grpcio-status==1.49.1
gym==0.23.0
gym-notices==0.0.8
h5py==3.7.0
idna==3.4
imageio==2.28.0
immutabledict==2.2.3
importlib-resources==5.9.0
install==1.3.5
ipykernel==6.17.1
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jax==0.3.15
jaxlib==0.3.15
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.1
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.3
jupyter_client==7.4.7
jupyter_core==5.1.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keras-nightly==2.13.0.dev2023020508
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mediapy==1.1.6
mistune==2.0.4
ml-collections==0.1.1
msgpack==1.0.4
mypy-protobuf==3.1.0
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
optax @ git+https://github.com/deepmind/optax.git@d9157132b8661e1453d156f591784ff27d6d1141
packaging==23.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.5.4
portpicker==1.5.2
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.33
proto-plus==1.22.1
protobuf==4.21.12
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.12
pybullet==3.2.5
pycparser==2.21
pygame==2.1.0
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rich==11.2.0
rlds==0.1.7
rsa==4.9
scenic @ git+https://github.com/google-research/scenic.git@ae21d9e884015aa7bc7cf1d489af53d16c249726
scipy==1.9.3
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
tb-nightly==2.12.0a20230205
tensorboard==2.11.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-addons==0.19.0
tensorflow-datasets==4.7.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-metadata==1.10.0
tensorflow-probability==0.17.0
tensorflow-text-nightly==2.12.0.dev20230203
tensorstore==0.1.22
termcolor==1.1.0
terminado==0.17.0
tf-agents==0.14.0
tf-estimator-nightly==2.13.0.dev2023020509
tf-hub-nightly==0.13.0.dev202302070818
tf-nightly==2.12.0.dev20230203
toml==0.10.2
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typeguard==2.13.3
types-protobuf==3.20.4
typing_extensions==4.3.0
urllib3==1.26.14
wcwidth==0.2.5
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
zipp==3.8.1
